Move flexmark KaTeX extension to less prominent location
