--- app/src/main/java/net/gsantner/markor/katex/internal/package-info.java
@@ -1,1 +1,0 @@
-package net.gsantner.markor.katex.internal;
