+++ app/src/main/java/net/gsantner/markor/util/flexmark/ext/katex/package-info.java
@@ -1,0 +1,1 @@
+package net.gsantner.markor.util.flexmark.ext.katex;
